feat: PromptBarButton
